Fix test name and import
